chore: satisfy strict clippy
Mechanical Clippy cleanup for the canonical Rust lint gate.

Validation:
- cargo fmt --all
- git diff --check
- cargo check
- cargo clippy --all-targets -- -D warnings
- cargo audit
- cargo test
- GitHub CI: fmt, check, clippy, test, bench-build, secrets-scan passed

+++ src/mcp/tools_term.rs
@@ -172,7 +172,7 @@ fn handle_term_read(args: &Value) -> ToolCallResult {
     let timeout_ms = args
         .get("timeout_ms")
         .and_then(|v| v.as_u64())
-        .unwrap_or(1000) as u64;
+        .unwrap_or(1000);
 
     let sessions = get_term_sessions();
     let mut guard = sessions.lock().unwrap();
